+++ Producer/RabbitMqProducerInterface.php
@@ -6,5 +6,11 @@ namespace MarfaTech\Bundle\RabbitQueueBundle\Producer;
 
 interface RabbitMqProducerInterface
 {
-    public function put(string $queueName, $data, array $options = [], string $routingKey = '');
+    public function put(
+        string $queueName,
+        $data,
+        array $options = [],
+        ?string $routingKey = null,
+        array $properties = []
+    );
 }