+++ Tests/Publisher/DeduplicatePublisherTest.php
@@ -14,8 +14,8 @@ use PhpAmqpLib\Message\AMQPMessage;
 
 class DeduplicatePublisherTest extends AbstractTestCase
 {
-    public const TEST_OPTIONS = ['key' => 'test'];
-    public const QUEUE_TYPE = QueueTypeEnum::FIFO | QueueTypeEnum::DEDUPLICATE;
+    protected const TEST_OPTIONS = ['key' => 'test'];
+    protected const QUEUE_TYPE = QueueTypeEnum::FIFO | QueueTypeEnum::DEDUPLICATE;
 
     /**
      * @throws HydratorNotFoundException
@@ -28,12 +28,12 @@ class DeduplicatePublisherTest extends AbstractTestCase
         $client = $this->createMock(RabbitMqClient::class);
         $client->expects(self::once())
             ->method('publish')
-            ->with(self::isInstanceOf(AMQPMessage::class), '', self::TEST_QUEUE_NAME)
+            ->with(self::getAmqpMockCallback(), '', self::TEST_QUEUE_NAME)
         ;
 
         $publisher = new DeduplicatePublisher($client, $hydratorRegistry, JsonHydrator::KEY);
 
-        $publisher->publish($definition, self::TEST_MESSAGE, self::TEST_OPTIONS);
+        $publisher->publish($definition, self::TEST_MESSAGE, self::TEST_OPTIONS, null, self::TEST_PARAMS);
 
         self::assertTrue(true);
     }